Обновление 12
Исправления перед добавлением SWAGGER

+++ src/main/java/com/example/testTask/dto/VacuumCleanerDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 public class VacuumCleanerDTO extends ModelDTO{
 
-    private Integer volume;
+    private Double volume;
     private Integer numberOfModes;
 }
