+++ src/main/java/com/example/testTask/dto/VacuumCleanerParams.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class VacuumCleanerParams extends ParamsDTO {
-    private Integer volume;
+    private Double volume;
     private Integer number_of_modes;
 }
